+++ src/elements/Session.php
@@ -611,21 +611,7 @@ class Session extends Element implements NestedElementInterface
         }
 
         if ($this->_purchasedSeatsCount === null) {
-            $query = (new Query())
-                ->from(['pt' => '{{%events_purchased_tickets}}'])
-                ->where(['pt.sessionId' => $this->id]);
-
-            if (Craft::$app->getDb()->columnExists('events_ticket_types', 'seatsPerTicket')) {
-                $query
-                    ->leftJoin(['tt' => '{{%events_ticket_types}}'], '[[pt.ticketTypeId]] = [[tt.id]]')
-                    ->select([new Expression('SUM(CASE WHEN COALESCE([[tt.seatsPerTicket]], 0) < 1 THEN 1 ELSE [[tt.seatsPerTicket]] END) AS total')]);
-            } else {
-                $query->select([new Expression('COUNT([[pt.id]]) AS total')]);
-            }
-
-            $total = $query->scalar();
-
-            $this->_purchasedSeatsCount = (int)($total ?? 0);
+            $this->_purchasedSeatsCount = Events::$plugin->getPurchasedTickets()->getPurchasedSeatsCountForSession($this->id);
         }
 
         return $this->_purchasedSeatsCount;